backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.20
sqlalchemy==2.0.36
bcrypt==4.2.1
pyjwt==2.10.1
python-dotenv==1.0.1
httpx==0.28.1
